fixed bug

+++ experiment_construction/learner_construction/tensorflow_learner.py
@@ -96,7 +96,7 @@ class TensorflowModel:
                 else:
                     gold_list.append(name)
             print(example["neighborhood"].entity_vertices.shape[0])
-            print("Swapping " + str(example["gold_entities"]) + " for " + str(gold_list))
+            print("Swapping " + str(example["gold_entities"]) + " for " + str(gold_list.astype(np.int32)))
 
             example["gold_entities"] = gold_list
             yield example
